Move some client sample chartUtils to analyses.utils
